requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sentence-transformers==3.0.1
faiss-cpu==1.12.0
langdetect==1.0.9
pydantic-settings==2.6.1
python-dotenv==1.0.1
numpy==2.1.2
httpx==0.27.2
pytest==8.3.3
python-multipart==0.0.9
